+++ src/Portenta_H7_AsyncTCP_Debug.h
@@ -9,12 +9,13 @@
   Built by Khoi Hoang https://github.com/khoih-prog/Portenta_H7_ASYNC_TCP
   Licensed under MIT license
  
-  Version: 1.1.0
+  Version: 1.2.0
   
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.0.0   K Hoang      06/10/2021 Initial coding for Portenta_H7 (STM32H7) with Vision-Shield Ethernet
   1.1.0   K Hoang      08/10/2021 Add support to Portenta_H7 (STM32H7) using Murata WiFi
+  1.2.0   K Hoang      15/10/2021 Workaround for issue with dns_gethostbyname not-working issue in mbed_portenta v2.5.2
  *****************************************************************************************************************************/
 
 #pragma once